encapsulated adminterminal into AdminModule.java
