+++ EasyAppointment/EasyAppointment-ejb/src/java/ejb/session/stateless/AdminEntitySessionBeanLocal.java
@@ -2,7 +2,7 @@ package ejb.session.stateless;
 
 import entity.AdminEntity;
 import util.exception.AdminNotFoundException;
-import util.exception.InvalidLoginException;
+import util.exception.InvalidLoginCredentialException;
 
 public interface AdminEntitySessionBeanLocal {
     public Long createAdminEntity(AdminEntity adminEntity);
@@ -11,5 +11,5 @@ public interface AdminEntitySessionBeanLocal {
 
     public AdminEntity retrieveAdminEntityByAdminEmail(String email) throws AdminNotFoundException;
 
-    public AdminEntity adminLogin(String email, String password) throws InvalidLoginException;
+    public AdminEntity adminLogin(String email, String password) throws InvalidLoginCredentialException;
 }
