+++ EasyAppointment/EasyAppointmentClient/src/java/easyappointmentclient/ServiceProviderModule.java
@@ -1,18 +1,9 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package easyappointmentclient;
 
 import ejb.session.stateless.ServiceProviderEntitySessionBeanRemote;
 import entity.ServiceProviderEntity;
 import java.util.Scanner;
 
-/**
- *
- * @author marcuslee
- */
 public class ServiceProviderModule {
     
     private ServiceProviderEntitySessionBeanRemote serviceProviderEntitySessionBeanRemote;
