+++ EasyAppointment/EasyAppointmentClient/src/java/easyappointmentclient/ServiceProviderTerminal.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package easyappointmentclient;
 
 import ejb.session.stateless.ServiceProviderEntitySessionBeanRemote;
@@ -10,10 +5,6 @@ import entity.ServiceProviderEntity;
 import java.util.Scanner;
 import util.exception.InvalidLoginCredentialException;
 
-/**
- *
- * @author marcuslee
- */
 public class ServiceProviderTerminal {
     
     private ServiceProviderEntitySessionBeanRemote serviceProviderEntitySessionBeanRemote;
