--- EasyAppointment/EasyAppointmentLibrary/src/util/exception/InvalidLoginException.java
@@ -1,14 +1,0 @@
-package util.exception;
-
-public class InvalidLoginException extends Exception 
-{
-    public InvalidLoginException()
-    {
-        
-    }
-    
-    public InvalidLoginException(String msg)
-    {
-        super(msg);
-    }
-}
